Added screens to go 3 levels of detail
Scan for devices
Query for Services
Get the characteristics of a service
Get the properties of a characteristic
We're now ready to read a characteristic and write to a characteristic, and get a response.
For example write, and get a file manifest
For write a file path and get the file contents.